Refactor: 이메일 인증 기능 분리 (#84)
